+++ src/FluentSerializer.Core.Tests/ObjectMother/SerializerContextExtensions.cs
@@ -1,3 +1,4 @@
+using FluentSerializer.Core.Configuration;
 using FluentSerializer.Core.Context;
 using FluentSerializer.Core.DataNodes;
 using FluentSerializer.Core.Naming.NamingStrategies;
@@ -10,6 +11,19 @@ namespace FluentSerializer.Core.Tests.ObjectMother;
 
 public static class SerializerContextMother
 {
+	/// <summary>
+	/// Create a minimal context mock for testing allowing for config changes
+	/// </summary>
+	public static Mock<TSerializer> UseConfig<TSerializer>(this Mock<TSerializer> contextMock, SerializerConfiguration configuration)
+		where TSerializer : class, ISerializer
+	{
+		contextMock
+			.Setup(context => context.Configuration)
+			.Returns(configuration);
+
+		return contextMock;
+	}
+
 	/// <summary>
 	/// Create a minimal context mock for testing
 	/// </summary>