+++ src/FluentSerializer.Core/Context/ISerializerCoreContext.cs
@@ -39,6 +39,16 @@ public interface ISerializerCoreContext
 	ISerializerCoreContext WithPathSegment(in PropertyInfo property);
 	/// <inheritdoc cref="WithPathSegment(in PropertyInfo)"/>
 	ISerializerCoreContext WithPathSegment(in Type type);
+
+	/// <summary>
+	/// Add an instance to the reference collection if not present already.
+	/// </summary>
+	bool TryAddReference(in object? instance);
+
+	/// <summary>
+	/// Check if a reference has been processed by the serializer already
+	/// </summary>
+	bool ContainsReference(in object? instance);
 }
 
 /// <summary>