+++ src/FluentSerializer.Json.Tests/Tests/Converting/Converters/ConvertibleConverterTests.cs
@@ -1,6 +1,7 @@
 using FluentAssertions;
 using FluentSerializer.Core.Context;
 using FluentSerializer.Core.Tests.ObjectMother;
+using FluentSerializer.Json.Configuration;
 using FluentSerializer.Json.Converting.Converters;
 using FluentSerializer.Json.DataNodes;
 using FluentSerializer.Json.Services;
@@ -26,7 +27,8 @@ public sealed class ConvertibleConverterTests
 	public ConvertibleConverterTests()
 	{
 		_sut = new ConvertibleConverter();
-		var serializerMock = new Mock<IAdvancedJsonSerializer>();
+		var serializerMock = new Mock<IAdvancedJsonSerializer>()
+			.UseConfig(JsonSerializerConfiguration.Default);
 		_contextMock = new Mock<ISerializerContext<IJsonNode>>()
 			.SetupDefault(serializerMock);
 	}