+++ src/FluentSerializer.Json.Tests/Tests/Converting/Converters/DateOnlyByFormatConverterTests.cs
@@ -2,6 +2,7 @@ using FluentAssertions;
 using FluentSerializer.Core.Context;
 using FluentSerializer.Core.Tests.ObjectMother;
 using FluentSerializer.Core.TestUtils.Extensions;
+using FluentSerializer.Json.Configuration;
 using FluentSerializer.Json.Converting.Converters;
 using FluentSerializer.Json.DataNodes;
 using FluentSerializer.Json.Services;
@@ -27,7 +28,8 @@ public sealed class DateOnlyByFormatConverterTests
 
 	public DateOnlyByFormatConverterTests()
 	{
-		var serializerMock = new Mock<IAdvancedJsonSerializer>();
+		var serializerMock = new Mock<IAdvancedJsonSerializer>()
+			.UseConfig(JsonSerializerConfiguration.Default);
 		_contextMock = new Mock<ISerializerContext<IJsonNode>>()
 			.SetupDefault(serializerMock);
 	}