+++ src/FluentSerializer.Json/Changelog.md
@@ -4,6 +4,7 @@
 
 - [#150](https://github.com/Marvin-Brouwer/FluentSerializer/issues/150) Added release notes
 - [#73](https://github.com/Marvin-Brouwer/FluentSerializer/issues/73) Added support for `DateTimeOffset`, `DateOnly`, `TimeOnly` and `TimeSpan`
+- [#101](https://github.com/Marvin-Brouwer/FluentSerializer/issues/101) Added reference loop detection
 
 ## v1.0.0